+++ buildSrc/src/main/kotlin/BuildConfig.kt
@@ -3,7 +3,7 @@ package com.unciv.build
 
 object BuildConfig {
     const val appName = "UncivCN"
-    const val appCodeNumber = 1248
-    const val appVersion = "4.21.6.4"
+    const val appCodeNumber = 1249
+    const val appVersion = "4.21.6.5"
     const val identifier = "com.unciv.app"
 }